requirements.txt
ffmpeg-python
transformers
ctranslate2
faster_whisper
scipy
streamlink
numpy
